requirements.txt
Django==2.1.5
django-summernote==0.8.11.4
djangorestframework==3.9.1
python-dotenv==0.10.1
pytz==2018.7
